Observe editors and assign completion provider in `zed`
