+++ crates/zed/src/main.rs
@@ -8,14 +8,18 @@ use backtrace::Backtrace;
 use chrono::Utc;
 use clap::{command, Parser};
 use cli::FORCE_CLI_MODE_ENV_VAR_NAME;
-use client::{parse_zed_link, Client, ClientSettings, DevServerToken, UserStore};
+use client::{
+    parse_zed_link, telemetry::Telemetry, Client, ClientSettings, DevServerToken, UserStore,
+};
 use collab_ui::channel_view::ChannelView;
+use copilot::Copilot;
+use copilot_ui::CopilotCompletionProvider;
 use db::kvp::KEY_VALUE_STORE;
-use editor::Editor;
+use editor::{Editor, EditorMode};
 use env_logger::Builder;
 use fs::RealFs;
 use futures::{future, StreamExt};
-use gpui::{App, AppContext, AsyncAppContext, Context, SemanticVersion, Task};
+use gpui::{App, AppContext, AsyncAppContext, Context, SemanticVersion, Task, ViewContext};
 use image_viewer;
 use isahc::{prelude::Configurable, Request};
 use language::LanguageRegistry;
@@ -176,6 +180,7 @@ fn main() {
             cx,
         );
         assistant::init(client.clone(), cx);
+        init_inline_completion_provider(client.telemetry().clone(), cx);
 
         extension::init(
             fs.clone(),
@@ -1067,3 +1072,18 @@ fn watch_file_types(fs: Arc<dyn fs::Fs>, cx: &mut AppContext) {
 
 #[cfg(not(debug_assertions))]
 fn watch_file_types(_fs: Arc<dyn fs::Fs>, _cx: &mut AppContext) {}
+
+fn init_inline_completion_provider(telemetry: Arc<Telemetry>, cx: &mut AppContext) {
+    if let Some(copilot) = Copilot::global(cx) {
+        cx.observe_new_views(move |editor: &mut Editor, cx: &mut ViewContext<Editor>| {
+            if editor.mode() == EditorMode::Full {
+                let provider = cx.new_model(|_| {
+                    CopilotCompletionProvider::new(copilot.clone())
+                        .with_telemetry(telemetry.clone())
+                });
+                editor.set_inline_completion_provider(provider, cx)
+            }
+        })
+        .detach();
+    }
+}
